Split off init tasks
Introduced templates/common.sh and templates/openvswitch_agent_init.sh
to run the container init tasks and run openvswitch_agent_init.sh with
the volume which stores the rendered configuration as parameter. The
init scripts are stored in its own configmap which allows us to mount
then with a different mask.

common.sh is intended to be a common functions lib for all the init
tasks for the neutron-operator.