Fix cmd/manager/main.go to be sdk 0.17 compatible
Right now building the neutron-operator fails with client-go error
as k8s.io/api and k8s.io/apimachinery are mapped to v0.18.2 which
is not compatible with go-client 0.17.4

Mapping k8s.io/api and k8s.io/apimachinery to v0.17.4 fixes the
go-client error, but need changes to cmd/manager/main.go for metrics.
Therefore we replace the main.go with a current one.