Fix defaulting PasswordSelectors
When a CRD has an optional struct field the defaulting of that field needs
special care. Both the field with the struct type need a full default value
defined and each individual subfields in the struct needs default value
defined. Otherwise in the first reconcile call the PasswordSelectors is
empty and then in the second reconcile it is filled.

Now documented via https://github.com/openstack-k8s-operators/docs/pull/12