Run neutronapi / db-sync as neutron:neutron
Now that we removed kolla bits from the startup path, we can unprivilege the
containers.

Note: since we still may mount HostPaths, we have to stick to hostmount-anyuid
for now. If this interface is revisited and e.g. switched to PVCs, then we
should be able to tighten SCC a bit more.