Reintroduce kolla into startup path for neutron-api
This is to keep neutron-api in line with [1] until we decide otherwise.

Keeping kolla has several drawbacks, specifically:

1. allowPrivilegeEscalation has to be true for sudo.
2. dropCapabilities cannot be ALL because kolla_set_configs failed to
   change to root gid.

This patch may be revisited later if we decide to tighten privileges
further by patching our kolla fork, or removing kolla completely from
the startup path. For now, we'll have to stick to common approach.

The discussion over kolla place can be continued in jira [2].

[1] https://github.com/openstack-k8s-operators/docs/blob/main/service_config.md
[2] https://issues.redhat.com/browse/OSPRH-2240