Hardcode mariadb database and account names to `neutron`
The latest mariadb-operator update made NewDatabase produce a database
and account named as neutronApi instance. But then, getDBWithName called
by CreateOrPatchDBByName expects account named as databaseUser. Which
breaks reconciliation.

While this may be a legit bug in mariadb, we can deal with it locally in
neutron by switching to NewDatabaseWithNamespace, which allows us to
hardcode the database and account CR names to "neutron", avoiding the
failing scenario.