Check Deployment ObservedGeneration to set DeploymentReadyCondition
DeploymentReadyCondition is something that depends on the number of
Replicas, but it should also be influenced by the ObservedGeneration. We
need to make sure that the readiness is set for the last CR version, and
this patch adds the additional check that we're rolling out to the other
operators as well.
ReadyCount is assigned only if we're looking at the last CR, and we
check and update the DeploymentReadyCondition only in this case.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>