Fix ObsGen condition
As per the agreement in [1], the controller should evaluate the current
CR readiness only when the Generation matches with the
ObservedGeneration. Any other edge case can be handled by a separate
check, but we do not want to risk to reach "ReadyCondition = True" when
we see a Generation "<" of the ObservedGeneration.

[1] https://github.com/openstack-k8s-operators/dev-docs/pull/102

Signed-off-by: Francesco Pantano <fpantano@redhat.com>