Add missing bits for DefaultConfigOverwrite
The input passed to DefaultConfigOverwrite was not
pushed to correct location where neutron api will use
those.

Also limited the use of this parameter by webhook just to provide
custom policy.yaml, can be extended if we need to support
other files later. It follows same patter as placement and
nova operator.

We already have "ExtraMounts" parameter which can be used
to push arbitary files/directory in neutron-api pod.

Closes-Issue: OSPRH-4298