Tune and enhance neutron-operator probes
This patch tunes the liveness and readiness probes to match other
operators [0][1][2][3].

We noticed that at times neutron tests would fail with 5XX errors,
indicating that the service was not avaiable during the execution of
the test, after investigation it was reported that neutron containers
were restarted due to probe failures.

Prior to this patch, neutron had the liveness and readiness probes set
to 5s where the majority of other operators had it set to something as
high as 30s (see links below).

This patch also removes the readiness probe from the neutron-api container
as it's redundant since it checks for readiness of the httpd container
(which is already checked by the httpd container readiness probe itself).

Resolves: https://issues.redhat.com/browse/OSPRH-9305

[0] https://github.com/openstack-k8s-operators/nova-operator/blob/27895863ccb7a5ef552e3d924d778fc7f664d056/pkg/novaapi/deployment.go#L49
[1] https://github.com/openstack-k8s-operators/keystone-operator/blob/1c577c9ef57612991d4139ab232eca61f0f43e83/pkg/keystone/deployment.go#L46
[2] https://github.com/openstack-k8s-operators/placement-operator/blob/d3a775bb32a963806949acd6556bd43ad901864d/pkg/placement/deployment.go#L41
[3] https://github.com/openstack-k8s-operators/glance-operator/blob/aa9b49b5810f76b4d06061191c5cb297eaa18ba8/pkg/glanceapi/statefulset.go#L62

Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>