Fix pagination links returned by neutron-server in non-TLS deployments
Neutron server returns links to the "next" and "previous" page of
results when pagination is requested by client.
In the neutron-httpd container configuration there was an issue that
"X-Forwarded-Proto" header was setting "https" protocol always, no
matter if TLS was enabled in the deployment or not. It caused issues
when TLS was disabled as client was doing request to the "http://"
endoint and got in response "https://" links to "next" and "prev" pages.

This patch fixes this issue by setting correct protocol for both cases:
TLS and non-TLS.

Closes-bug: #10632