Make apache timeout configurable
Added a parameter 'apiTimeout' to allow customization
to the Apache timeout.
The default is set to 120s which we do set for
HAProxy timeouts currently.

To be able to change the HAProxy value based on the `apiTimeout` with
any update (and not just the first time) the code adds a custom
annotation "api.neutron.openstack.org/timeout" with the value that was
initially set, this way flags it as being set by the neutron-operator.

There will be follow up patch in openstack-operator
to utilize the method 'SetDefaultRouteAnnotations' to set
these default route annotations.

Resolves: OSPRH-10843