Extend CRDs to support a RabbitMqCluster reference as AMQP instance for notifications.
To provide a way for human user to define the notifications bus for
Neutronapi CR via the OpenStackControlPlane CR's neutron section, add a new
field to the NeutronAPISpecCore as that is the struct used when the
OpenStackControlPlaneSpec struct is defined.

At NeutronSpec level having just "" or a name of the bus is OK as either
we need a bus, or we don't need notifications at all.

     “” - disabled,
     "foo" - external RabbitMQ Bus

Assume that in the implementation the required information will be
transferred via the existing internal secret via notifications specific
transport_url key (as the transport_url key is already taken by the RPC
bus value).